Codigo_fonte/requirements.txt
pandas
numpy
scikit-fuzzy
scikit-surprise
nltk
fuzzywuzzy
python-Levenshtein
requests
time
os
re 
python
